[Utils] Fixed the user-defined literals finally

+++ Runtime/engine/utils/alias.h
@@ -34,8 +34,8 @@ namespace polos
 
     using cstring = const char*;
 
-    inline constexpr float  operator""_ms(long double time) { return 0.001f * time; }
-    inline constexpr float  operator""_us(long double time) { return 0.001f * 0.001f * time; }
+    inline constexpr float operator""_ms(long double time) { return 0.001f * static_cast<float>(time); }
+    inline constexpr float operator""_us(long double time) { return 0.001f * 0.001f * static_cast<float>(time); }
 }
 
 #endif /* POLOS_CORE_UTILS_ALIAS_H_ */
